requirements.txt
kafka-python
confluent_kafka
elasticsearch==7.13
kibana-api
sqllineage
acryl-datahub==0.11.0